Important Update
In this change a reset of the lists "userList" and "solutionList" is performed at the end of each cycle of the main loop.

This is a must when more than 1 homework is being worked on in the loop, which is quite likely.

These changes are mandatory and are just an oversight during development. Therefor I ship them in a new branch.